responsive main page
