+++ src/components/ui/navbar/StickyNavbar.jsx
@@ -1,0 +1,96 @@
+import React, { useState, useEffect, useCallback } from "react";
+import { useNavigate } from "react-router-dom";
+import NavbarHeader from "./NavbarHeader";
+import WalletModal from "../../wallet/WalletModal";
+import { useDispatch } from "react-redux";
+import { logoutSuccess } from "../../../actions/authActions";
+import { API } from "../../../config";
+import axios from "axios";
+import ERC20Contract from "../../../contract/ERC20Contract";
+import { CategoryBar } from "../CategoryBar";
+import { debounce } from "lodash";
+
+export default function StickyNavbar({ children, isLoggedIn }) {
+  const [openNav, setOpenNav] = useState(false);
+  const [openModal, setOpenModal] = useState(false);
+  const [walletAddress, setWalletAddress] = useState(null);
+  const [balance, setBalance] = useState(null);
+
+  const navigate = useNavigate();
+  const dispatch = useDispatch();
+
+  const handleOpen = useCallback(async () => {
+    try {
+      const token = localStorage.getItem("token");
+      const response = await axios.get(`${API.GETWALLETADDRESS}`, {
+        headers: {
+          "Content-Type": "application/json",
+          Authorization: `Bearer ${token}`,
+        },
+      });
+      const address = response.data.existingWallet.wallet_address;
+      setWalletAddress(address);
+      await getBalance(address);
+      setOpenModal((cur) => !cur);
+    } catch (error) {
+      console.error("Error fetching data:", error);
+      navigate("/login");
+    }
+  }, [navigate]);
+
+  const handleSignOut = useCallback(() => {
+    localStorage.removeItem("token");
+    dispatch(logoutSuccess());
+    navigate("/");
+  }, [dispatch, navigate]);
+
+  useEffect(() => {
+    const handleResize = () => window.innerWidth >= 960 && setOpenNav(false);
+    window.addEventListener("resize", handleResize);
+    return () => {
+      window.removeEventListener("resize", handleResize);
+    };
+  }, []);
+
+  const getBalance = useCallback(async (address) => {
+    if (!address) {
+      console.error("Invalid wallet address");
+      return;
+    }
+    try {
+      const erc20Contract = await ERC20Contract.getInstance();
+      const result = await erc20Contract.balanceOf(address);
+      setBalance(result.toString());
+    } catch (error) {
+      console.error("Failed to get balance:", error);
+    }
+  }, []);
+
+  const debouncedNavigate = useCallback(
+    debounce((path) => navigate(path), 300),
+    [navigate]
+  );
+
+  return (
+    <div className="grid min-h-[140px] w-full overflow-x-scroll rounded-lg p-6 lg:overflow-visible">
+      <div className="-m-6 max-h-screen overflow-scroll">
+        <NavbarHeader
+          openNav={openNav}
+          setOpenNav={setOpenNav}
+          handleOpen={handleOpen}
+          handleSignOut={handleSignOut}
+          isLoggedIn={isLoggedIn}
+          debouncedNavigate={debouncedNavigate}
+        />
+        <CategoryBar />
+        {children}
+        <WalletModal
+          open={openModal}
+          handleOpen={handleOpen}
+          address={walletAddress}
+          initialBalance={balance}
+        />
+      </div>
+    </div>
+  );
+}
