+++ src/pages/MainPage.jsx
@@ -3,7 +3,7 @@ import { Typography } from "@material-tailwind/react";
 import bgBlurWebP from "../assets/bg-blur1.webp";
 import bgBlur2WebP from "../assets/bg-blur2.webp";
 
-const StickyNavbar = lazy(() => import("../components/ui/StickyNavbar"));
+const StickyNavbar = lazy(() => import("../components/ui/navbar/StickyNavbar"));
 const CarouselWithContent = lazy(() =>
   import("../components/ui/CarouselWithContent")
 );
@@ -12,6 +12,8 @@ const EcommerceCard = lazy(() => import("../components/ui/EcommerceCard"));
 const FeatureBlock = lazy(() => import("../components/ui/FeatureBlock"));
 
 const MainPage = ({ isLoggedIn }) => {
+  const token = localStorage.getItem("token");
+  console.log(token);
   const products = useMemo(
     () => [
       {
@@ -46,11 +48,11 @@ const MainPage = ({ isLoggedIn }) => {
             backgroundRepeat: "no-repeat, no-repeat",
             backgroundPosition: "-100% 70%, 130% 100%",
           }}
-          className="py-8 h-[680px]"
+          className="py-8 h-auto lg:h-[680px] bg-cover bg-no-repeat"
         >
-          <div className="mx-40 px-4">
-            <div className="flex justify-between">
-              <div className="w-7/12">
+          <div className="mx-4 lg:mx-40 px-4">
+            <div className="flex flex-col lg:flex-row xl:justify-between">
+              <div className="w-full xl:w-7/12">
                 <Typography variant="h5" className="mb-5 ml-3 text-gray-700">
                   Featured Project
                 </Typography>
@@ -61,11 +63,11 @@ const MainPage = ({ isLoggedIn }) => {
                   status={70}
                 />
               </div>
-              <div className="w-5/12 ml-5">
+              <div className="w-full xl:w-5/12 xl:ml-5 mt-4 xl:mt-0">
                 <Typography variant="h5" className="mb-5 ml-1 text-gray-700">
                   Recommended for you
                 </Typography>
-                <div className="grid grid-cols-1 sm:grid-cols-1 md:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 2xl:grid-cols-2 gap-4">
+                <div className="grid grid-cols-1 sm:grid-cols-2 gap-4">
                   {products.map((product, index) => (
                     <EcommerceCard
                       key={index}
@@ -89,10 +91,13 @@ const MainPage = ({ isLoggedIn }) => {
             backgroundRepeat: "no-repeat",
             backgroundPosition: "40% 100%",
           }}
+          className="bg-cover bg-no-repeat py-8"
         >
-          <FeatureBlock />
+          <div className="mx-4 lg:mx-40">
+            <FeatureBlock />
+          </div>
         </div>
-        <div className="mx-40">
+        <div className="container mx-auto px-4">
           <FooterWithLogo />
         </div>
       </StickyNavbar>
